LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-8-expected.txt

هذه جملة باللغة الإنجليزية: The quick brown fox jumped over the lazy dog.

Verifies that the selection extends to select the contents of the second line, and the ending selection handle is placed on the left.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isVisuallyContiguous is true
PASS lastLineBounds.left - finalEndHandlePoint.x is <= widthOfPeriod + 4
PASS getSelection().toString().includes('The quick brown fox jumped over the lazy dog') is true
PASS successfullyParsed is true

TEST COMPLETE
